lib/mito/io: move reader and writer in summit and vtk folder respectively
